// src/interfaces/Imovies.ts
export interface movieResonse {
    page: number,
    results: Movie[],
    total_pages: number,
    total_results: number
}

export interface GenresResponse {
genres: genre[];
}

export interface genre {
    id: number,
    name: string,
}

export interface Movie {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    genres: genre[],
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}

export interface ProductionCompany {
    id: number;
    logo_path: string | null;
    name: string;
  }
  
  export interface ProductionCountry {
    iso_3166_1: string;
    name: string;
  }
  
  export interface SpokenLanguage {
    english_name: string;
    iso_639_1: string;
    name: string;
  }
  
  export interface MovieDetail {
    adult: boolean;
    backdrop_path: string;
    belongs_to_collection: null | {
      id: number;
      name: string;
      poster_path: string;
      backdrop_path: string;
    };
    budget: number;
    genres: genre[];
    homepage: string;
    id: number;
    imdb_id: string;
    origin_country: string[];
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    production_companies: ProductionCompany[];
    production_countries: ProductionCountry[];
    release_date: string;
    revenue: number;
    runtime: number;
    spoken_languages: SpokenLanguage[];
    status: string;
    tagline: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
  }